Configure pipelines for building and publishing cli images
Finished configuring the pipelines for publishing cli docker images.
Decided to have a script which will be in charge of flying the pipelines
as new major versions are added.

The pipeline ingests the branch associated with the major version and
expects the dockerfile to be located under the `docker`
folder in the cli repo.

Co-authored-by: Dominic Roberts <dominicr@vmware.com>